requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
apache-airflow==2.7.1